LPS-12839
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@63011 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-service/src/com/liferay/portal/kernel/util/StringUtil.java
@@ -1372,10 +1372,10 @@ public class StringUtil {
 
 	public static String toHexString(Object obj) {
 		if (obj instanceof Integer) {
-			return toHexString((Integer)obj);
+			return toHexString(((Integer)obj).intValue());
 		}
 		else if (obj instanceof Long) {
-			return toHexString((Long)obj);
+			return toHexString(((Long)obj).longValue());
 		}
 		else {
 			return String.valueOf(obj);
